Fix integer kind to make gcc9 happy
Gcc 9 complains about different integer kinds in the iand function.
According to the standard, this is also not allowed, so we should fix
it. Up to now, we were just lucky that the compilers did not complain.

+++ src/td/pes_mask_out_inc.F90
@@ -778,7 +778,7 @@ subroutine pes_mask_output_full_mapM(pesK, file, Lk, ll, how, sb, pmesh)
   character(len=*),  intent(in) :: file
   FLOAT,             intent(in) :: Lk(:,:)
   integer,           intent(in) :: ll(:)  
-  integer,           intent(in) :: how
+  integer(8),        intent(in) :: how
   type(simul_box_t), intent(in) :: sb 
   FLOAT, optional,   intent(in) :: pmesh(:,:,:,:)
